pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vaultnotes"
version = "0.1.0"
description = "Publish selected Obsidian vault folders as a browsable GitHub Pages site."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Maleen Kidiwela" }]
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
vaultnotes = "vaultnotes.cli:main"

[project.urls]
Homepage = "https://github.com/MaleenKidiwela/vaultnotes"

[tool.hatch.build.targets.wheel]
packages = ["src/vaultnotes"]

[tool.hatch.build.targets.wheel.force-include]
"src/vaultnotes/templates" = "vaultnotes/templates"
